+++ database/java/migrations/ticket-order-db-migrations/ticket-order-transactional-migrations/src/main/resources/db/repeatable/R__event_orders.sql
@@ -2,7 +2,7 @@ CREATE OR REPLACE VIEW event_orders AS
 SELECT
     event_order_id,
     event_id,
-    customer_reference,
+    customer_id,
     row_number,
     place_number,
     place_type,